requirements.txt
azure-cosmos==4.6.0
azure-functions==1.17.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
numpy==1.26.2
requests==2.31.0
tiktoken==0.8.0
tenacity==8.2.3
langchain==0.2.10
langchain-community==0.2.10
langchain-core==0.2.25
numexpr==2.10.0
langgraph==0.0.64
pandas==2.2.2
langchain_experimental==0.0.63
tabulate===0.9.0
stripe==10.5.0
tavily-python==0.5.0
google_search_results==2.4.2
langchain-openai < 0.2.5
weasyprint==63.1
python-dotenv
openpyxl